requirements_enhanced.txt
# TRD-UQ Enhanced Requirements
# Core dependencies for uncertainty quantification and temporal analysis

# Machine Learning & Deep Learning
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0

# Uncertainty Quantification
scikit-learn>=1.0.0        # For calibration and metrics
scipy>=1.7.0               # Statistical functions

# Configuration & Serialization
PyYAML>=6.0                # For configuration files

# System Monitoring
psutil>=5.8.0              # System performance monitoring

# Data Analysis & Visualization
matplotlib>=3.5.0          # For uncertainty visualization
seaborn>=0.11.0            # Enhanced statistical plots

# Database
mysql-connector-python>=8.0.0

# Utilities
numpy>=1.21.0
opencv-python>=4.5.0
tqdm>=4.60.0               # Progress bars for training